requirements.txt
certifi==2017.7.27.1
chardet==3.0.4
Django==1.10.5
django-cors-middleware==1.3.1
django-oauth-toolkit==1.0.0
djangorestframework==3.6.3
idna==2.5
oauthlib==2.0.2
requests==2.18.2
urllib3==1.22
